tests/requirements.txt
pytest
pytest-asyncio
PyNaCl